+++ socialapp/cmd/cli/commands/delete/role/role.go
@@ -76,7 +76,7 @@ func GetCmd() *cli.Command {
 
 			apiClient := client.NewAPIClient(cfg)
 
-			httpResponse, err := apiClient.RoleAPI.DeleteRole(ctx, int32(roleID)).Execute()
+			httpResponse, err := apiClient.RoleAPI.DeleteRole(ctx, roleID).Execute()
 			if err != nil {
 				bodyBytes, _ := io.ReadAll(httpResponse.Body)
 				fmt.Printf("Error: %v\nResponse: %s\n", err, string(bodyBytes))
