+++ socialapp/cmd/cli/commands/delete/scope/scope.go
@@ -76,7 +76,7 @@ func GetCmd() *cli.Command {
 
 			apiClient := client.NewAPIClient(cfg)
 
-			httpResponse, err := apiClient.ScopeAPI.DeleteScope(ctx, int32(scopeID)).Execute()
+			httpResponse, err := apiClient.ScopeAPI.DeleteScope(ctx, scopeID).Execute()
 			if err != nil {
 				bodyBytes, _ := io.ReadAll(httpResponse.Body)
 				fmt.Printf("Error: %v\nResponse: %s\n", err, string(bodyBytes))
